Factor out generic SignaturesReport code into WeeklyReportable concern
