+++ app/reports/signatures_report.rb
@@ -1,28 +1,7 @@
-require 'csv'
+require_relative 'concerns/weekly_reportable'
 
 class SignaturesReport
-  attr_reader :begin, :end
-
-  ##
-  # @param begin_date [String,Date]
-  #   Date or String parseable as a Date
-  # @param end_date [String,Date]
-  #   Date or String parseable as a Date
-  def initialize(begin_date, end_date)
-    @begin = Date.parse(begin_date.to_s)
-    @end = Date.parse(end_date.to_s)
-  end
-
-  ##
-  # @param ending [String,Date]
-  #   Date or String parseable as a Date
-  #
-  # @return [SignaturesReport]
-  def self.for_week(ending:)
-    end_date = Date.parse(ending.to_s)
-    begin_date = end_date - 6
-    new(begin_date, end_date)
-  end
+  include WeeklyReportable
 
   ##
   # @private
@@ -45,14 +24,4 @@ class SignaturesReport
                       name
                       email
                       signature).freeze
-
-  def to_csv
-    join = pull_join
-    CSV.generate(write_headers: false, headers: JOINED_HEADERS) do |csv|
-      # Don't want to rely on `write_headers: true` since we want still
-      # header row in the CSV file even when there is no data.
-      csv << JOINED_HEADERS
-      join.each { |record| csv << record.attributes }
-    end
-  end
 end
